Modulize server.py
