+++ requirements.txt
@@ -1,14 +1,6 @@
-Flask==0.10.1
 Flask-Cors==2.1.2
 Flask-SocketIO==2.0
-Jinja2==2.8
-MarkupSafe==0.23
-Werkzeug==0.11.3
-eventlet==0.18.1
-greenlet==0.4.9
-itsdangerous==0.24
+Flask==0.10.1
+eventlet==0.18.3
 python-engineio==0.8.7
 python-socketio==1.0
-six==1.10.0
-wheel==0.24.0
-wsgiref==0.1.2
